PCI: Drop pcibios_pm_ops from the PCI subsystem
The "struct dev_pm_ops pcibios_pm_ops", declared in include/linux/pci.h and
defined in drivers/pci/pci-driver.c, provided arch-specific hooks when a
PCI device was doing a hibernate transisiton.

Although it lost its last usage after
394216275c7d ("s390: remove broken hibernate / power management support")
patch.

After that, instances of it are found only in drivers/pci/pci-driver.c and
include/linux/pci.h, which are now unnecessary. Thus it is safe and
reasonable to remove even that.

Reported-by: Bjorn Helgaas <helgaas@kernel.org>
Signed-off-by: Vaibhav Gupta <vaibhavgupta40@gmail.com>